requirements.txt
python-telegram-bot
groq
pyodbc
python-dotenv